requirements.txt
streamlit==1.12.0
openai==0.11.0
requests==2.27.1
python-decouple==3.6
PyMuPDF==1.18.19
pdfplumber==0.5.28
python-dotenv==0.19.2
pyyaml==5.4.1
Pillow==8.4.0
